+++ tests/unit/system.c
@@ -23,54 +23,54 @@ void test_reset_cause() {
     for (i = 0; i < 4; i++) {
         MCUSR = 1 << i;
 
-        cause = system_get_reset_status();
+        cause = hal_system_get_reset_status();
 
         TEST_ASSERT_EQUAL(1 << i, cause);
         TEST_ASSERT_EQUAL(0, MCUSR);
 
         if (i == 0) {
-            TEST_ASSERT_EQUAL(system_power_on_reset, cause);
+            TEST_ASSERT_EQUAL(hal_system_power_on_reset, cause);
         } else if (i == 1) {
-            TEST_ASSERT_EQUAL(system_external_reset, cause);
+            TEST_ASSERT_EQUAL(hal_system_external_reset, cause);
         } else if (i == 2) {
-            TEST_ASSERT_EQUAL(system_brownout_reset, cause);
+            TEST_ASSERT_EQUAL(hal_system_brownout_reset, cause);
         } else if (i == 3) {
-            TEST_ASSERT_EQUAL(system_watchdog_reset, cause);
+            TEST_ASSERT_EQUAL(hal_system_watchdog_reset, cause);
         }
     }
 }
 
 void test_set_modes() {
-    struct system_watchdog_t config;
+    struct hal_system_watchdog_t config;
 
-    config.mode = system_watchdog_disabled;
-    system_set_watchdog(config);
+    config.mode = hal_system_watchdog_disabled;
+    hal_system_set_watchdog(config);
     TEST_ASSERT_EQUAL(0, WDTCSR & BIT(WDIE));
     TEST_ASSERT_EQUAL(0, WDTCSR & BIT(WDE));
 
-    config.mode = system_watchdog_interrupt_mode;
-    system_set_watchdog(config);
+    config.mode = hal_system_watchdog_interrupt_mode;
+    hal_system_set_watchdog(config);
     TEST_ASSERT_EQUAL(BIT(WDIE), WDTCSR & BIT(WDIE));
     TEST_ASSERT_EQUAL(0, WDTCSR & BIT(WDE));
 
-    config.mode = system_watchdog_reset_mode;
-    system_set_watchdog(config);
+    config.mode = hal_system_watchdog_reset_mode;
+    hal_system_set_watchdog(config);
     TEST_ASSERT_EQUAL(0, WDTCSR & BIT(WDIE));
     TEST_ASSERT_EQUAL(BIT(WDE), WDTCSR & BIT(WDE));
 
-    config.mode = system_watchdog_interrupt_and_reset_mode;
-    system_set_watchdog(config);
+    config.mode = hal_system_watchdog_interrupt_and_reset_mode;
+    hal_system_set_watchdog(config);
     TEST_ASSERT_EQUAL(BIT(WDIE), WDTCSR & BIT(WDIE));
     TEST_ASSERT_EQUAL(BIT(WDE), WDTCSR & BIT(WDE));
 }
 
 void test_set_cycles() {
     uint8_t i;
-    struct system_watchdog_t config;
+    struct hal_system_watchdog_t config;
 
     for (i = 0; i < 10; i++) {
-        config.cycles = system_watchdog_2k_cycles + i;
-        system_set_watchdog(config);
+        config.cycles = hal_system_watchdog_2k_cycles + i;
+        hal_system_set_watchdog(config);
         TEST_ASSERT_EQUAL(config.cycles, WDTCSR & 0b1111);
     }
 }